trace_processor: add track_id column to slice table
This CL adds the track_id column to the slice table; this column
is always filled and contains the track id associated with the slice.

This CL also changes the slice tracker to use the track id instead
of ref + ref_type to figure out the nesting of tracks.

Context: go/perfetto-tp-refactor
Bug: 135177627

+++ src/trace_processor/storage_columns.h
@@ -334,7 +334,7 @@ class StringVectorAccessor : public Accessor<NullTermStringView> {
   }
 
   NullTermStringView Get(uint32_t idx) const override {
-    const char* ptr = (*string_map_)[(*deque_)[idx]];
+    const char* ptr = (*string_map_)[static_cast<size_t>((*deque_)[idx])];
     return ptr ? NullTermStringView(ptr) : NullTermStringView();
   }
